function-invoker: Clean up runtime download after install
Previously the runtime package tarball was left behind in the runtime
layer after the package was installed, so ended up in the built image.

Now the file is downloaded to a temporary directory instead of the layer,
and cleaned up after the npm install.

The `local runtime_package` statement being on a separate line is required due to:
https://github.com/koalaman/shellcheck/wiki/SC2155#problematic-code-in-the-case-of-local

Closes GUS-W-9238585.