Allow app to change Node Versions (#77)
It was noted that if an app tries to deploy using Node 16 and then tries to deploy the same app using Node 14 it fails. It turns out it will fail for any node version change and not just a downgrade in version.

```
ERROR: failed to build: Near line 10 (last key parsed ''): Key 'metadata' has already been defined.
```

The reason for this error is that the nodejs.toml file contains metadata key twice which makes it invalid tool and pack cannot process it. Here's what the contents:

```
build = true
launch = true
cache = true

[metadata]
 version = "16.3.0"
cache = true
build = true
launch = true
[metadata]
version = "14.17.0"
```

The fix is to write over the toml file instead of appending to it. I'm worried though there may be other places with a similar bug such as installing yarn. 

There are Shpec tests and we would in theory be able to unit test this specific function to make sure this bug doesn't happen. The existing shpec tests take awhile to run though and it doesn't look like there a way to "focus" them. 

I think it's a good idea to have at least some basic integration tests with cutlass on the node function buildpack, so I think this is still worthwhile, but we might want to consider adding a unit test for this, and using unit tests for similar issues. Also it might be worth while to split up unit tests for more focused  (faster) executions.