yarn: Make Cutlass test use the test metabuildpack
Since currently it's using the standard (non-test) metabuildpack, which
causes failures when the versions are bumped in #87.

I was not able to use the `:default` alias (to avoid specifying the default
test metabuildpack again directly),  since the current `spec_helper.rb` teardown
doesn't work correctly with multiple tests.

Follow-up to #86.

[skip changelog]

Refs GUS-W-9481665.