Tests: Run Cutlass tests with the Heroku-20 based builder image
Previously the non-functions Node Cutlass test was using the Heroku-18
based builder image instead, which is not the default on the platform.

In addition, I've fixed the variable name in `spec_helper.rb` since this file
is for the non-functions test, not the functions test. This was presumably
a copy-paste error when creating the tests.

Closes GUS-W-9483311.