Make publishing script fail if duplicate buildpack IDs found
Previously if duplicate buildpack IDs existed in the repository, the publish
script would publish the first buildpack found, and exit 0.

The Node.js test buildpacks have the same IDs as the real buildpacks, leading
to the test buildpacks being published instead of the intended buildpack.

With this change, the publish script now traverses the repository and errors
if more than one buildpack with the same ID was found.

In addition, I've added a bit more logging around the name/path/version of
the found buildpack, which would have aided debugging this issue.

This PR is the Node.js equivalent of:
https://github.com/heroku/buildpacks-jvm/pull/129

Refs GUS-W-9672962.