Use a unique buildpack ID for the test nodejs-function buildpack
Currently the nodejs-function test meta-buildpack has the same buildpack
ID as the actual nodejs-function meta-buildpack, causing the publishing
script to publish it instead of the intended buildpack.

The publishing script is being updated in another PR to fail if duplicate buildpack
IDs are found, so we must give the test buildpack a unique ID.

The new name suffix matches the style used for the test nodejs (non-function)
meta-buildpack and also all of the JVM test meta-buildpacks.

I've also changed the test meta-buildpack version to `0.0.0` for parity with the
others.

Fixes GUS-W-9672962.