fix: use the proper command path for local/global runtime
This is a fix against the changes introduced with [#373](https://github.com/heroku/buildpacks-nodejs/pull/373) where the use of `npx` was not detecting the runtime command (`sf-fx-runtime-nodejs`) already installed globally.

The effect of which was the runtime to be downloaded when the Function application initially starts up which is not ideal since it causes slower boot times + may cause non-deterministic behavior.

[W-11854639](https://gus.lightning.force.com/lightning/r/ADM_Work__c/a07EE000017vsgyYAA/view?ws=%2Flightning%2Fr%2FADM_Epic__c%2Fa3QEE000000MuDF2A0%2Fview)