Adjust method for skipping check changelog for Dependabot PRs
This means the manual `labels` attributes can be removed,
allowing Dependabot's default labels to take effect (which are
more informative, since it applies a label to reflect the language
etc). This approach matches that used in our other repos.

Also ports the permissions and step naming tweaks applied to the
check changelog check in our other repos.