Fix cutlass integration tests
This fixes the cutlass integration tests which no longer need to use a copy of each meta-buildpack with local dependency references as paths to produce a testable buildpack. These can now use the actual meta-buildpacks and rely on `cargo libcnb package` and `libcnb:{buildpack_id}` references to produce the same result.