Support yarn 2.4.3
When listing the upstream yarn distributions there is a version `2.4.3` which is tagged as `berry` in npm for the `yarn` package (https://www.npmjs.com/package/yarn/v/2.4.3).

Our mirroring automation expects yarn versions > 1.x to be distributed under the `@yarnpkg/cli-dist` but there is no published package for `@yarnpkg/cli-dist@2.4.3`.

This change should fix our Yarn mirroring which appears broken in GitHub Action runs due to version `2.4.3` belonging to the `yarn`, not `@yarnpkg/cli-dist`.