Update buildpack metadata in buildpack.toml (#692)
Adjusts the buildpack `name`, `description` and `keywords` in the
`buildpack.toml` of all buildpacks to match the style discussed in:
https://github.com/heroku/cnb-builder-images/issues/408

These fields are used by the CNB registry and can also be seen in the
output of `pack builder inspect`. They are not used by `pack build` or
Kodon. In addition to standardising the metadata, one of the objectives
here is to ensure users don't accidentally use component buildpacks when
they really wanted the composite buildpack.

The `keywords` field has intentionally been omitted from the functions
buildpacks since they are EOL and so don't want to be made more
discoverable. (The CNB registry doesn't currently ingest/display the
keywords currently, but may in the future.)

Lastly, I've cleaned up the `[metadata]` table/subtables to match the
concise style used by the Procfile and Python CNBs.

GUS-W-14121598.