Update the legacy NPM buildpack to Buildpack API 0.7 (#721)
Since Buildpack API 0.6 has been deprecated for some time, and is no
longer supported by newer versions  of `lifecycle`, which is causing
CI to fail since the integration tests in this repo use `heroku/builder:22`,
which has just had its lifecycle version upgraded:
https://github.com/heroku/cnb-builder-images/pull/431

These CI failures are not reflective of production (which is working fine),
since Functions in production are actually run via the functions builder
image (and not `heroku/builder:22`), which was intentionally left on the
old lifecycle version. However, this functions builder image is not available
on Docker Hub, so can't be used by the integration tests in this repo.

Improving parity of the integration tests here might be worth considering
longer term, however, upgrading the Buildpack API version was a simpler
(and arguably correct thing to do regardless) step for now.

There are no breaking changes that affect this buildpack in the new API
version, so no buildpack changes were required:
https://github.com/buildpacks/spec/releases/tag/buildpack%2Fv0.7

Fixes #720.
GUS-W-14493124.