Restrict metrics layer installation to Node >=14.10
The main functionality of this PR is to modify `nodejs-engine` so that the metrics script is only include when the Node.js version is 14.10 or higher + integration tests to verify that behavior.

This PR also includes some refactoring around code related to obtaining the current Node.js version which is used by both `nodejs-engine` and `nodejs-npm-engine`. This is now available in `nodejs-utils` as `node::get_node_version`.

Some common error handling output was also moved into `nodejs-utils`.

[W-14838686](https://gus.lightning.force.com/lightning/r/a07EE00001iPnjsYAC/view)