Update CODEOWNERS to reduce review notification noise
This modification to the `CODEOWNERS` file is based off suggestions from @edmorley to help reduce review notification overload for the team.

For most PRs, the Languages team will be requested for review except for automation-based changes which are assigned to specific language owners since they often require further action such as performing a release, etc.