Fix/regenerate Cargo.lock
Dependabot is currently failing since it doesn't like the fact that the
version of the Ruby `commons` module (which is imported via a
GitHub URL) has changed from `0.1.0` to `0.0.0` after:
https://github.com/heroku/buildpacks-ruby/pull/248

To fix this the lockfile entry for `commons` needs regenerating.

However, whilst I'm here I've just regenerated the whole lockfile,
to pick up any other in-range version updates.