Fix flaky integration tests
Several of the metrics tests are temporal-based and, while we can get decent reliability by tweaking the timeouts used, it's not a guarantee that the chosen timeouts will always be effective which can lead to unwanted flakiness during test runs or test runs that wait longer than necessary.

To balance the tradeoff between wait time and test execution time, this PR introduces `wait_for(condition, timeout)` which will continually poll the state of the given `condition` until it passes or the `timeout` is exceeded. This should help to minimize the wait time required to guarantee the expected outcome.