Extend timeouts for test reliability (#919)
* Extend timeouts for test reliability

The tests in this PR fail occasionally in CI test runs and then pass on re-run. I'm extending the timeouts in the hopes this will result in more stability.

* Extend timeouts for test reliability

The tests in this PR fail occasionally in CI test runs and then pass on re-run. I'm extending the timeouts in the hopes this will result in more stability.