Fixed bug in `nodejs-engine` layer configuration
As part of the migration to the new layer struct API (#880), the layer handler function names for `attach_runtime_metrics` and `configure_web_env` got reversed.

This should have been caught by integration tests but the tests that would exercise that area can be flaky due to timing issues. On further inspection, I realized that the intermittent failures I've seen recently weren't due to timing but because the `wait_for` function should only be used for checking positive signals (`assert_contains!`) and not negative signals (`assert_not_contains!`).

I've added extra logging so that these positive signals can be captured before checking for the absence of a message.

This PR also includes a small lint fix.