Remove check and warning for `node_modules` directory
This warning was carried over from the classic Heroku Node.js buildpack which did support both `npm install` and `npm ci`. If a `node_modules` directory existed and there was no lockfile present, this could cause issues with the dependencies installed.

This buildpack only supports `npm ci` so this check and warning is not required. If the `node_modules` directory is already present, `npm ci` will automatically remove it.