Add retries to binary downloads
Swapped out the usage of `download_file` from `libherokubuildpack` which does not currently perform retries for a combination of `reqwest` + retry middleware. Except for a couple of error messages, this is mostly a drop-in replacement.

One drawback here is that an async runtime (`tokio`) is required to make this work. These buildpacks were written without async in mind, so the `download_file_sync` function added in this PR hides most of this away. This isn't the most optimal way to handle this, but due to where the downloads happen, it shouldn't be an issue.