Ensure top-level `env` gets updated when layers are created
When these CNBs were separate, this wasn't necessary because the lifecycle would take care of updating the environment after each CNB executed. Now that we're in a singular execution context, the `env` passed through to commands needs to be updated when layers are created so all our expected commands are available on the path.