Remove multiple lockfile/missing lockfile error handling
This error case needs to be handled entirely differently now. It should be much easier to raise this check earlier in the process (i.e.; before install Node.js) now that all tooling choices are happening in the same context. For now though, this check will need to be removed.