Replace npm inventory with packument for version resolution
This PR replaces the inventory-based npm version resolution strategy with an equivalent method that uses the same process npm itself uses.

For example, if you run `npm info npm@latest`, a request will be made to `https://registry.npmjs.com/npm` to fetch what npm calls a `Packument`. The `Packument` is in JSON format and contains metadata about all published versions of a package under the `versions` field. This `versions` field is essentially a richer version of what our inventory file contained as each entry has the package version, distribution url, checksums, etc.

While it also could have been possible to replace npm version resolution with a call to `npm info npm@{requested_version}`, the advantage of requesting the `Packument` directly is that the buildpack can cache this information for use in subsequent builds. The cached `Packument` will be invalidated whenever a new version of `npm` is published.