Replace Yarn inventory with packument for version resolution
This PR replaces the inventory-based Yarn version resolution strategy with an equivalent method that uses the same process npm itself uses for version resolution.

For example, if you run `npm info yarn@latest`, a request will be made to `https://registry.npmjs.com/yarn` to fetch what npm calls a `Packument`. The `Packument` is in JSON format and contains metadata about all published versions of a package under the `versions` field. This `versions` field is essentially a richer version of what our inventory file contained as each entry has the package version, distribution url, checksums, etc.

While it also could have been possible to replace Yarn version resolution with a call to `npm info yarn@{requested_version}`, the advantage of requesting the `Packument` directly is that the buildpack can cache this information for use in subsequent builds. The cached `Packument` will be invalidated whenever a new version of `yarn` is published.

> [!NOTE]
> There are actually two packages for Yarn:
> - `yarn` for v1 (and earlier)
> - `@yarnpkg/cli-dist` for v2 (and later)
> The appropriate packument will be fetched based on what the user has requested for the Yarn version.