Replace npm inventory with packument (#1121)
* Replace npm inventory with packument for version resolution

This PR replaces the inventory-based npm version resolution strategy with an equivalent method that uses the same process npm itself uses.

For example, if you run `npm info npm@latest`, a request will be made to `https://registry.npmjs.com/npm` to fetch what npm calls a `Packument`. The `Packument` is in JSON format and contains metadata about all published versions of a package under the `versions` field. This `versions` field is essentially a richer version of what our inventory file contained as each entry has the package version, distribution url, checksums, etc.

While it also could have been possible to replace npm version resolution with a call to `npm info npm@{requested_version}`, the advantage of requesting the `Packument` directly is that the buildpack can cache this information for use in subsequent builds. The cached `Packument` will be invalidated whenever a new version of `npm` is published.

* Update all snapshots to show new request status

* Update CHANGELOG.md

Signed-off-by: Colin Casey <casey.colin@gmail.com>

* Fix unit test

* Apply suggestions from code review

Co-authored-by: Richard Schneeman <richard.schneeman+foo@gmail.com>
Signed-off-by: Colin Casey <casey.colin@gmail.com>

* Add metadata serialization test

* Add error message

* Invalid the cached layer if the packument cannot be parsed

---------

Signed-off-by: Colin Casey <casey.colin@gmail.com>
Co-authored-by: Richard Schneeman <richard.schneeman+foo@gmail.com>