Demo: Buildpack Config
This is an experiment with controlling buildpack functionality using a tiered configuration system. The idea is that a buildpack can provide a public configuration schema for configuration values that can be set in a unified manner by:
- buildpack operators
- buildpack authors
- end users

Let's take the example of setting the version of Node.js to install. A buildpack (in this demo, `heroku/nodejs-engine`) could source this from several locations (in order or priority):
1. `package.json` - the `engines.node` field is the standard location to set this information in a Node.js project & it would be controllable by the end user.
2. `project.toml` - using a configuration namespace of `[com.heroku.buildpacks.nodejs]`, the end user could set the `runtime.version` field.
3. Buildplan - Using `requires` metadata for the configuration namespace of `com.heroku.buildpacks.nodejs`, a contributing buildpack could set the `runtime.version` field.

Not included in this demo, but this idea could easily be extended to source from `ENV` using a similiar namespacing process (e.g.; `COM_HEROKU_BUILDPACKS_NODEJS_RUNTIME_VERSION=20.x`).