Restructure individual buildpacks to use inner detect/build
This is the most significant code change involved in this refactor. For each of the individual buildpacks we need to:
- Remove their pre-existing Buildpack implementations
- Repurpose their `detect` functions to be called from within the `build` context of the top-level buildpack and make these return boolean results instead of detect results
- Ensure that the top-level `Env` and `BuildResultBuilder` are woven through each `build` function

The top-level buildpack then uses these functions to recreate the ordering that used to be part of the `heroku/nodejs` composite buildpack and execute only the logic that used to be done by the detected buildpacks.

This works in conjunction with the top-level buildpack's code which ensures that the previous buildpack ord