Refactor individual buildpacks into one singular buildpack (#1169)
* Relocate project files and remove composite CNB

Each of the individual CNBs need the bulk of their code moved into the `nodejs` CNB which will contain the same functionality under a singular buildpack.

This also means the `meta-buildpacks/nodejs` composite CNB can be removed as it will be converted into our single, component CNB.

* Add project structures to satisfy Cargo workspace loading

Also includes stubbed out `main` methods for the CNBs that will deprecated in the initial v5 release.

* Add `mod.rs` files for each buildpack area

* Add required libraries

* Replace `use crate::{...}` bindinds with `use super::{...}` to match relative locations of modules

* Move `yarn/yarn.rs` code into `yarn/main.rs` since the file name interferes with the module name

* Fix imports, normalize use of `BuildpackBuildContext`, `BuildpackError`, and `BuildpackResult<T>`

These alias types will help to convert everything to the same context now that each individual buildpack no longer runs in isolation. This also involved fixing up some of the `From<ErrorType> for BuildpackError` conversions which now need to roll up into the top-level `NodeJsBuildpackError`.

Omitted from this commit are the changes required for `build`, `detect`, and `error` handling.

* Update error handling routines

* Fix compile errors for deprecated buildpacks

* Restructure individual buildpacks to use inner detect/build

This is the most significant code change involved in this refactor. For each of the individual buildpacks we need to:
- Remove their pre-existing Buildpack implementations
- Repurpose their `detect` functions to be called from within the `build` context of the top-level buildpack and make these return boolean results instead of detect results
- Ensure that the top-level `Env` and `BuildResultBuilder` are woven through each `build` function

The top-level buildpack then uses these functions to recreate the ordering that used to be part of the `heroku/nodejs` composite buildpack and execute only the logic that used to be done by the detected buildpacks.

This works in conjunction with the top-level buildpack's code which ensures that the previous buildpack ord

* Clippy and check fixes

* Update error snapshots with new buildpack name

* Ensure env is being passed through to npm install commands

* Drop application lockfiles check as this needs to be handled differently in the single buildpack setup

* Rename `node_build_scripts` buildplan id to `heroku/nodejs`

* Rename Yarn installation layer since it conflicts with `dist` which is already used by the Node installation layer from `node-engine`

* Expand the snapshot filter to capture `heroku/nodejs`

* Some cleanup of the top-level entrypoint

* Updated snapshots passing tests

* Add deprecation errors to detect + tests

* Refactor main logic to align better with how previous order detection worked

* Update with changes from #1166 and #1167

* Update CHANGELOG.md with links to previous buildpacks

* Move buildpack code to top-level `src` and `tests` directories

* Update README.md

* Fix CHANGELOG links

* Update inventory workflow paths

* Ensure root buildpack gets named properly in CI

* Surround jq expression in parenthesis to fix syntax error