Rebuild snapshots on inventory update (#1180)
When new Node.js releases are detected by automation and the inventory file gets updated, this causes several snapshots to become invalidated and integration tests will fail. This adds an extra annoying step to reviewing a new Node.js release because the dev must checkout, rebuild the snapshots, and commit the changes for the automated PR to pass status checks.

- Uses [cargo-nextest](https://nexte.st/) as the integration test runner since it has retry support.
- Uses SHA for `actions/checkout` and `actions/create-github-app-token` versions.
- Adds steps required to compile and run integration tests
- Bumps the slow test warning to 5m after more local testing