Refactor inventory update task
Refactors the binary for updating the inventory into its own standalone crate and adds an alias so it can be easily invoked via `cargo update-nodejs-inventory`.

The primary benefit is that it will allow removing the existing binary located in `nodejs-utils` + its dependencies and code which are unrelated to the buildpack code. This will be done in a follow-up PR.

Some other extra benefits added in this refactoring:
- Stricter argument parsing from `clap` for enum values
- GPG verification of Node.js release checksum file signature before adding assets to the inventory to ensure the integrity of the checksums we validate against.