Switch Node.js install to use full streaming download/extract/verify
The previous version of this operated using distinct steps for the download, extract, and verification of the Node.js download. This involves creating a bunch of intermediate files, reading the full contents of the download archive multiple times (for checksum and extract), and extracting to a temporary location so that pathing could be fixed before copying to the final destination directory.

The new download functionality leverages async functionality to stream directly from the http response to both a hash verifier + gzip/tar extractor which can handle the path rewriting to the destination directly.