Move npm install & remove `npm_engine`
This moves the npm installation layer into the top-level modules. With this final change, the `npm_engine` can be removed.

One important detail to note is that the integration test fixtures did have to be corrected here as the way the previous npm install process executed would modify the node `dist` layer. The npm install is entirely isolated now. This is why the layer details from `pack` reflect this change for the `dist` and `npm_engine` layers.