This builds on top of #1199. The goal of this PR is to: - eliminate the contents of the former `pnpm-engine` CNB found in `src/pnpm_engine` - leverage the functions from `util::npm_registry` for downloading and installing packages from the registry to replace the previous Corepack-based functionality
This simplified process for package manager installation has laid the groundwork for moving away from using the `corepack` tooling directly while still supporting the use of the `packageManager` field in `package.json`. This is especially important as Corepack won't be officially supported/bundled with Node.js in the future.
Because of this fundamental change, the integration test snapshots do show both layer name and buildpack output changes.