Support installation of Yarn via `yarn.engines` and `packageManager` from `package.json`
This uses the same packument installation process laid out in earlier commits which already handles `npm` and `pnpm` installs. This works without modifications to the `src/yarn` CNB code due to how it checks if Yarn is installed and skips its own internal install process in that case.

With this change, all Corepack-related code can finally be removed.