Detect vendored Yarn installation
The prior Yarn CNB would bootstrap Yarn 1.x to deal with projects using a vendored Yarn distribution. This isn't necessary as the file referenced in the `yarnPath` of `.yarnrc.yml` should be a fully functional `yarn` binary. This PR introduces a new uncached layer for this case that creates a symlink from the `yarnPath` to `yarn` and makes this available at both build and launch.