Reintegrate npm install (#1205)
* Switch to use `BuildpackConfig` for `npm_install`

The `buildpack_config` module is intended to replace the individual handling of buildplans and project.toml as configuration sources. Eventually this module will replace `utils::buildplan` and `utils::config` but the old `pnpm_install` and `yarn` CNB flows still rely on them.

* Change `determine_package_manager` to return a non-optional value

Since npm is bundled with Node.js, we can default to that when no package manager information is provided.

* Chain runtime installation flow

This mirrors the package manager install flow

* Remove `npm_install` detect

* Lift dependency installation from `npm_install` into top-level

* Lift build script execution from `npm_install` into top-level

* Lift dev dependency pruning from `npm_install` into top-level

* Move npm runtime configuration into an exec.d script

* Lift process registration from `npm_install` into top-level

* Merge npm-related integration tests

* Reintegrate yarn (#1206)

* Remove `yarn` detect

* Switch to use `BuildpackConfig` for `yarn`

* Lift global cache setup from `yarn` into top-level

* Lift dependency installation from `yarn` into top-level

For consistency with the npm cache layer, the Yarn cache layer has been renamed to `yarn_cache`. This can be seen in changes to the integration test snapshots.

* Lift build script execution from `yarn` into top-level

* Lift dev dependency pruning from `yarn` into top-level

* Lift process registration from `yarn` into top-level

This final step allows the `src/yarn` directory to finally be removed.

* Reintegrate pnpm install (#1208)

* Remove `pnpm_install` detect

* Switch to use `BuildpackConfig` for `pnpm_install`

* Lift dependency installation from `pnpm_install` into top-level

* Add pnpm persistent store cleanup

This feature requires the use of the CNB store to keep a counter of how many builds the cache has been around for.

* Lift build script execution from `pnpm_install` into top-level

* Lift dev dependency pruning from `pnpm_install` into top-level

* Lift process registration from `pnpm_install` into top-level

This final step allows the `src/pnpm_install` directory to finally be removed.

* Make `ErrorMessage` the primary Err-type for the buildpack

* Merge pnpm integration tests